Fix collision detection

+++ src/visgraph.cpp
@@ -358,7 +358,7 @@ namespace visgraph
     {
         // See https://www.geeksforgeeks.org/orientation-3-ordered-points/
         // for details of below formula.
-        int val = (q.y - p.y) * (r.x - q.x) -
+        double val = (q.y - p.y) * (r.x - q.x) -
                   (q.x - p.x) * (r.y - q.y);
 
         if (val == 0)
